+++ assets/js/data/cart/index.ts
@@ -15,8 +15,9 @@ import reducer, { State } from './reducers';
 import { controls as sharedControls } from '../shared-controls';
 import { controls } from './controls';
 import type { SelectFromMap, DispatchFromMap } from '../mapped-types';
+import { pushChanges } from './push-changes';
 
-registerStore< State >( STORE_KEY, {
+const registeredStore = registerStore< State >( STORE_KEY, {
 	reducer,
 	actions,
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -24,6 +25,8 @@ registerStore< State >( STORE_KEY, {
 	selectors,
 	resolvers,
 } );
+
+registeredStore.subscribe( pushChanges );
 
 export const CART_STORE_KEY = STORE_KEY;
